#76 reorganize go to location dialog
